Cleaning up glTF examples
