--- Assets/MixedRealityToolkit.Examples/Demos/Gltf/Models/BoomBox.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a0b03050fab2bec4e98cb00211ed9b10
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
