--- Assets/MixedRealityToolkit.Examples/Demos/Gltf/Models/BoomBox/README.md.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: f302864003ffaad4cafe16aa41537a94
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
