--- Assets/MixedRealityToolkit.Examples/Demos/Gltf/Models/BoomBox/glTF-Binary.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 207aa6f5c788df84f84853ba619db3a6
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
