--- Assets/MixedRealityToolkit.Examples/Demos/Gltf/Models/BoomBox/glTF.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 1227be416f12ac945aab749e1665337a
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
