--- Assets/MixedRealityToolkit.Examples/Demos/Gltf/Models/BoomBox/glTF/BoomBox.bin.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 56704c4d7391c9341aa6ac2c4a13fdef
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
