--- Assets/MixedRealityToolkit.Examples/Demos/Gltf/Models/BoomBox/glTF/BoomBox_Mat.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6241d9e26428c9c46bd6461733516190
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
