+++ Assets/MixedRealityToolkit.Examples/Demos/Gltf/Models/Duck.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 82250c53def6b994c8e7a85a2acc5cff
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
