+++ Assets/MixedRealityToolkit.Examples/Demos/Gltf/Models/Duck/README.md.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 744c50b64bc67af42a0e774a6dd180cf
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
