+++ Assets/MixedRealityToolkit.Examples/Demos/Gltf/Models/Duck/glTF.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c2a8cb69a96a78c40b41f49c2a493352
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
