+++ Assets/MixedRealityToolkit.Examples/Demos/Gltf/Models/Duck/glTF/Duck.gltf.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: 998d845703c61f748b6e659d91f1ea56
+ScriptedImporter:
+  fileIDToRecycleName:
+    100000: main/glTF Scene Duck
+    100002: main/glTF Scene Duck/glTF Node 0
+    100004: main/glTF Scene Duck/glTF Node 0/glTF Node 1
+    100006: main/glTF Scene Duck/glTF Node 0/glTF Node 2
+    400000: main/glTF Scene Duck/Transform
+    400002: main/glTF Scene Duck/glTF Node 0/Transform
+    400004: main/glTF Scene Duck/glTF Node 0/glTF Node 1/Transform
+    400006: main/glTF Scene Duck/glTF Node 0/glTF Node 2/Transform
+    2300000: main/glTF Scene Duck/glTF Node 0/glTF Node 2/MeshRenderer
+    3300000: main/glTF Scene Duck/glTF Node 0/glTF Node 2/MeshFilter
+    4300000: LOD3spShape
+    11400000: glTF data
+    100100000: main
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: 
+  script: {fileID: 11500000, guid: 8f83316e73954ae4fa56b79705605b45, type: 3}
