+++ Assets/MixedRealityToolkit.Examples/Demos/Gltf/Models/Duck/glTF/Duck0.bin.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 6f578a7215b4e20438f60466058e49cb
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
