+++ Assets/MixedRealityToolkit.Examples/Demos/Gltf/Models/Duck/glTF/DuckMaterial.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2ba2ae3ddcf29c249a77fdae0ead8369
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
