--- Assets/MixedRealityToolkit.Examples/Demos/Gltf/Models/Lantern.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8025f1ef53cc3ce42aa211a9b115ed8b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
