--- Assets/MixedRealityToolkit.Examples/Demos/Gltf/Models/Lantern/README.md
@@ -1,9 +1,0 @@
-# Lantern
-## Screenshot
-
-![screenshot](../../../../../../Documentation/Images/gltf/lantern_screenshot.jpg)
-
-## License Information
-
-[![CC0](http://i.creativecommons.org/p/zero/1.0/88x31.png)](http://creativecommons.org/publicdomain/zero/1.0/)  
-To the extent possible under law, Microsoft has waived all copyright and related or neighboring rights to this asset.
