--- Assets/MixedRealityToolkit.Examples/Demos/Gltf/Models/Lantern/README.md.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 5a57a28a88264c44598a228d4e9dbfb0
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
