--- Assets/MixedRealityToolkit.Examples/Demos/Gltf/Models/Lantern/glTF-Binary.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: cccf19254717f5a46b61e444b8e304d1
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
