--- Assets/MixedRealityToolkit.Examples/Demos/Gltf/Models/Lantern/glTF-Binary/Lantern.glb.meta
@@ -1,35 +1,0 @@
-fileFormatVersion: 2
-guid: 8f1cdfa5cfbaa244d96e41b67f7a440e
-ScriptedImporter:
-  fileIDToRecycleName:
-    100000: main/glTF Scene Lantern
-    100002: main/glTF Scene Lantern/Lantern
-    100004: main/glTF Scene Lantern/Lantern/LanternPole_Body
-    100006: main/glTF Scene Lantern/Lantern/LanternPole_Chain
-    100008: main/glTF Scene Lantern/Lantern/LanternPole_Lantern
-    400000: main/glTF Scene Lantern/Transform
-    400002: main/glTF Scene Lantern/Lantern/Transform
-    400004: main/glTF Scene Lantern/Lantern/LanternPole_Body/Transform
-    400006: main/glTF Scene Lantern/Lantern/LanternPole_Chain/Transform
-    400008: main/glTF Scene Lantern/Lantern/LanternPole_Lantern/Transform
-    2100000: LanternPost_Mat
-    2300000: main/glTF Scene Lantern/Lantern/LanternPole_Body/MeshRenderer
-    2300002: main/glTF Scene Lantern/Lantern/LanternPole_Chain/MeshRenderer
-    2300004: main/glTF Scene Lantern/Lantern/LanternPole_Lantern/MeshRenderer
-    2800000: Texture_0
-    2800002: Texture_1
-    2800004: Texture_2
-    2800006: Texture_3
-    3300000: main/glTF Scene Lantern/Lantern/LanternPole_Body/MeshFilter
-    3300002: main/glTF Scene Lantern/Lantern/LanternPole_Chain/MeshFilter
-    3300004: main/glTF Scene Lantern/Lantern/LanternPole_Lantern/MeshFilter
-    4300000: LanternPole_Body
-    4300002: LanternPole_Chain
-    4300004: LanternPole_Lantern
-    11400000: glTF data
-    100100000: main
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-  script: {fileID: 11500000, guid: 75911c45f303c8f45927f5efeaa50cc4, type: 3}
