--- Assets/MixedRealityToolkit.Examples/Demos/Gltf/Models/Lantern/glTF.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e9df3f2122823324bb39004383b08d26
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
