--- Assets/MixedRealityToolkit.Examples/Demos/Gltf/Models/Lantern/glTF/Lantern.bin.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 97b55e0fbe107b74fba508653770d1a4
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
