--- Assets/MixedRealityToolkit.Examples/Demos/Gltf/Models/Lantern/glTF/Lantern.gltf
@@ -1,360 +1,0 @@
-{
-  "accessors": [
-    {
-      "bufferView": 0,
-      "componentType": 5126,
-      "count": 926,
-      "type": "VEC2"
-    },
-    {
-      "bufferView": 1,
-      "componentType": 5126,
-      "count": 926,
-      "type": "VEC3"
-    },
-    {
-      "bufferView": 2,
-      "componentType": 5126,
-      "count": 926,
-      "type": "VEC4"
-    },
-    {
-      "bufferView": 3,
-      "componentType": 5126,
-      "count": 926,
-      "type": "VEC3",
-      "max": [
-        7.74559927,
-        12.8321095,
-        2.31570983
-      ],
-      "min": [
-        -7.74559927,
-        -12.8321095,
-        -2.31570983
-      ]
-    },
-    {
-      "bufferView": 4,
-      "componentType": 5123,
-      "count": 2616,
-      "type": "SCALAR"
-    },
-    {
-      "bufferView": 5,
-      "componentType": 5126,
-      "count": 756,
-      "type": "VEC2"
-    },
-    {
-      "bufferView": 6,
-      "componentType": 5126,
-      "count": 756,
-      "type": "VEC3"
-    },
-    {
-      "bufferView": 7,
-      "componentType": 5126,
-      "count": 756,
-      "type": "VEC4"
-    },
-    {
-      "bufferView": 8,
-      "componentType": 5126,
-      "count": 756,
-      "type": "VEC3",
-      "max": [
-        0.129208073,
-        0.6523504,
-        0.129208073
-      ],
-      "min": [
-        -0.129208073,
-        -0.6523504,
-        -0.129208073
-      ]
-    },
-    {
-      "bufferView": 9,
-      "componentType": 5123,
-      "count": 3744,
-      "type": "SCALAR"
-    },
-    {
-      "bufferView": 10,
-      "componentType": 5126,
-      "count": 2463,
-      "type": "VEC2"
-    },
-    {
-      "bufferView": 11,
-      "componentType": 5126,
-      "count": 2463,
-      "type": "VEC3"
-    },
-    {
-      "bufferView": 12,
-      "componentType": 5126,
-      "count": 2463,
-      "type": "VEC4"
-    },
-    {
-      "bufferView": 13,
-      "componentType": 5126,
-      "count": 2463,
-      "type": "VEC3",
-      "max": [
-        1.03408229,
-        2.529281,
-        1.03408468
-      ],
-      "min": [
-        -1.03408229,
-        -2.529281,
-        -1.03408468
-      ]
-    },
-    {
-      "bufferView": 14,
-      "componentType": 5123,
-      "count": 9822,
-      "type": "SCALAR"
-    }
-  ],
-  "asset": {
-    "generator": "glTF Tools for Unity",
-    "version": "2.0"
-  },
-  "bufferViews": [
-    {
-      "buffer": 0,
-      "byteLength": 7408
-    },
-    {
-      "buffer": 0,
-      "byteOffset": 7408,
-      "byteLength": 11112
-    },
-    {
-      "buffer": 0,
-      "byteOffset": 18520,
-      "byteLength": 14816
-    },
-    {
-      "buffer": 0,
-      "byteOffset": 33336,
-      "byteLength": 11112
-    },
-    {
-      "buffer": 0,
-      "byteOffset": 44448,
-      "byteLength": 5232
-    },
-    {
-      "buffer": 0,
-      "byteOffset": 49680,
-      "byteLength": 6048
-    },
-    {
-      "buffer": 0,
-      "byteOffset": 55728,
-      "byteLength": 9072
-    },
-    {
-      "buffer": 0,
-      "byteOffset": 64800,
-      "byteLength": 12096
-    },
-    {
-      "buffer": 0,
-      "byteOffset": 76896,
-      "byteLength": 9072
-    },
-    {
-      "buffer": 0,
-      "byteOffset": 85968,
-      "byteLength": 7488
-    },
-    {
-      "buffer": 0,
-      "byteOffset": 93456,
-      "byteLength": 19704
-    },
-    {
-      "buffer": 0,
-      "byteOffset": 113160,
-      "byteLength": 29556
-    },
-    {
-      "buffer": 0,
-      "byteOffset": 142716,
-      "byteLength": 39408
-    },
-    {
-      "buffer": 0,
-      "byteOffset": 182124,
-      "byteLength": 29556
-    },
-    {
-      "buffer": 0,
-      "byteOffset": 211680,
-      "byteLength": 19644
-    }
-  ],
-  "buffers": [
-    {
-      "uri": "Lantern.bin",
-      "byteLength": 231324
-    }
-  ],
-  "images": [
-    {
-      "uri": "Lantern_baseColor.png"
-    },
-    {
-      "uri": "Lantern_roughnessMetallic.png"
-    },
-    {
-      "uri": "Lantern_normal.png"
-    },
-    {
-      "uri": "Lantern_emissive.png"
-    }
-  ],
-  "meshes": [
-    {
-      "primitives": [
-        {
-          "attributes": {
-            "TEXCOORD_0": 0,
-            "NORMAL": 1,
-            "TANGENT": 2,
-            "POSITION": 3
-          },
-          "indices": 4,
-          "material": 0
-        }
-      ],
-      "name": "LanternPole_Body"
-    },
-    {
-      "primitives": [
-        {
-          "attributes": {
-            "TEXCOORD_0": 5,
-            "NORMAL": 6,
-            "TANGENT": 7,
-            "POSITION": 8
-          },
-          "indices": 9,
-          "material": 0
-        }
-      ],
-      "name": "LanternPole_Chain"
-    },
-    {
-      "primitives": [
-        {
-          "attributes": {
-            "TEXCOORD_0": 10,
-            "NORMAL": 11,
-            "TANGENT": 12,
-            "POSITION": 13
-          },
-          "indices": 14,
-          "material": 0
-        }
-      ],
-      "name": "LanternPole_Lantern"
-    }
-  ],
-  "materials": [
-    {
-      "pbrMetallicRoughness": {
-        "baseColorTexture": {
-          "index": 0
-        },
-        "metallicRoughnessTexture": {
-          "index": 1
-        }
-      },
-      "normalTexture": {
-        "index": 2
-      },
-      "emissiveFactor": [
-        1.0,
-        1.0,
-        1.0
-      ],
-      "emissiveTexture": {
-        "index": 3
-      },
-      "name": "LanternPost_Mat"
-    }
-  ],
-  "nodes": [
-    {
-      "mesh": 0,
-      "translation": [
-        -3.82315421,
-        13.01603,
-        0.0
-      ],
-      "name": "LanternPole_Body"
-    },
-    {
-      "mesh": 1,
-      "translation": [
-        -9.582001,
-        21.0378723,
-        0.0
-      ],
-      "name": "LanternPole_Chain"
-    },
-    {
-      "mesh": 2,
-      "translation": [
-        -9.582007,
-        18.0091515,
-        0.0
-      ],
-      "name": "LanternPole_Lantern"
-    },
-    {
-      "children": [
-        0,
-        1,
-        2
-      ],
-      "rotation": [
-        0.0,
-        1.0,
-        0.0,
-        0.0
-      ],
-      "name": "Lantern"
-    }
-  ],
-  "scene": 0,
-  "scenes": [
-    {
-      "nodes": [
-        3
-      ]
-    }
-  ],
-  "textures": [
-    {
-      "source": 0
-    },
-    {
-      "source": 1
-    },
-    {
-      "source": 2
-    },
-    {
-      "source": 3
-    }
-  ]
-}
