--- Assets/MixedRealityToolkit.Examples/Demos/Gltf/Models/Lantern/glTF/LanternPost_Mat.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6c40f43a0d2811f4aadb3fac33244a32
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
