+++ Assets/MixedRealityToolkit.Examples/Demos/Gltf/Scripts/Editor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3a742c66f8043de46b9f1458b9f99963
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
