+++ Assets/MixedRealityToolkit.Examples/Demos/Gltf/Scripts/Editor/TestGltfLoadingEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f60f682de8202ae4da594a9b895a64fa
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
